requirements.txt
asgiref==3.8.1
boto3==1.38.18
botocore==1.38.18
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
